Added separate pid_mapping files
Keeping pid calc files separate to allow general use of pid.
